+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   `ActionItemAddress` component
+
 ## [0.2.13] - 2023-08-31
 
 ### Fixed