+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.14] - 2023-09-12
+
 ### Added
 
 -   `ActionItemAddress` component